Migrate Query.clientStats
